refactor: [EL-4914] consolidate tour constants and update imports for interactive tours
