+++ src/[fsd]/features/interactive-tours/lib/constants/agentTour.constants.js
