+++ src/[fsd]/features/interactive-tours/lib/constants/index.js
@@ -1,1 +1,4 @@
 export { CARD_WIDTH_PX } from './interactiveTour.constants';
+export { AGENT_TOUR_COMPLETION, AGENT_TOUR_ID, agentTourSteps } from './agentTour.constants';
+export { CHAT_TOUR_COMPLETION, CHAT_TOUR_ID, chatTourSteps } from './chatTour.constants';
+export { SIDEBAR_TOUR_COMPLETION, SIDEBAR_TOUR_ID, sidebarTourSteps } from './sidebarTour.constants';
