+++ src/[fsd]/features/interactive-tours/lib/constants/sidebarTour.constants.js
