+++ src/[fsd]/features/interactive-tours/lib/hooks/useInteractiveTourController.hooks.js
@@ -1,15 +1,13 @@
 import { useCallback, useMemo, useReducer } from 'react';
 
-import { AGENT_TOUR_ID } from '../agentTour';
-import { CHAT_TOUR_ID } from '../chatTour';
+import { AGENT_TOUR_ID, CHAT_TOUR_ID, SIDEBAR_TOUR_ID } from '../constants';
 import { initialState, lsCompletedKey, lsPromptKey, tourReducer } from '../helpers';
-import { SIDEBAR_TOUR_ID } from '../sidebarTour';
 
 // ─── Tour loaders (lazy) ───────────────────────────────────────────────────────
 const TOUR_LOADERS = {
-  [CHAT_TOUR_ID]: () => import('../chatTour').then(m => m.chatTourSteps),
-  [AGENT_TOUR_ID]: () => import('../agentTour').then(m => m.agentTourSteps),
-  [SIDEBAR_TOUR_ID]: () => import('../sidebarTour').then(m => m.sidebarTourSteps),
+  [CHAT_TOUR_ID]: () => import('../constants/chatTour.constants').then(m => m.chatTourSteps),
+  [AGENT_TOUR_ID]: () => import('../constants/agentTour.constants').then(m => m.agentTourSteps),
+  [SIDEBAR_TOUR_ID]: () => import('../constants/sidebarTour.constants').then(m => m.sidebarTourSteps),
 };
 
 /**
